enable xai detection e2e tests
